+++ core/indigo-core/molecule/src/molecule_substructure_matcher.cpp
@@ -1438,10 +1438,10 @@ int MoleculeSubstructureMatcher::_compare_frequency_asc(BaseMolecule& mol, int i
     int label2 = mol.getAtomNumber(i2);
     int idx1, idx2;
 
-    for (idx1 = 0; idx1 < (int)NELEM(labels_by_freq); idx1++)
+    for (idx1 = 0; idx1 < NELEM(labels_by_freq); idx1++)
         if (label1 == labels_by_freq[idx1])
             break;
-    for (idx2 = 0; idx2 < (int)NELEM(labels_by_freq); idx2++)
+    for (idx2 = 0; idx2 < NELEM(labels_by_freq); idx2++)
         if (label2 == labels_by_freq[idx2])
             break;
